+++ src/templates/prepared-test/test-templates/usecase.ts
@@ -1,0 +1,4 @@
+export default `describe('{{name}}', () => {
+  {{#if errorExample}}{{>Error errorExample name=name }}{{/if}}
+  {{#if successExample}}{{>Success successExample name=name }}{{/if}}
+});`;